Convert the admin nav select into an Admin::Combobox component
The admin navbar's page picker was a selectize `<select>` navigated by a
jQuery handler in the vendored bundle, beside an "All" link guarded by a
helper. Both now live in Admin::Combobox::Component, rendering
UI::Forms::Combobox::Component with a Stimulus controller that navigates on
selection.

AdminHelper loses the two nav link tables and the two predicates that read
them; what's left is admin_search_filtered?, which admin/bikes/index also
uses to decide whether to show sitewide counts.
