+++ app/views/admin/bikes/index.html.erb
@@ -252,7 +252,7 @@
   <%= number_display(Ownership.where(current: true, claimed: true).count) %>
   claimed, <%= number_display(PublicImage.count) %> bike images
   <em>(<%= number_display(PublicImage.where("created_at >= ?", Time.current.beginning_of_day).count) %> today)</em>
-<% end unless admin_nav_display_view_all %>
+<% end unless admin_search_filtered? %>
 
 <% table_view = capture do %>
   <%= render Admin::BikesTable::Component.new(bikes: @bikes, render_sortable: true, render_multi_check: @multi_delete) %>
