+++ app/views/layouts/admin.html.erb
@@ -82,18 +82,7 @@
             </li>
           <% end %>
         </ul>
-        <% if admin_nav_display_view_all %>
-          <%= link_to admin_nav_select_link_active[:path], class: "nav-link text-muted em" do %>
-            All
-            <div class="d-none d-sm-inline-block d-lg-none d-xl-inline-block">
-              <%# Remove "Config:" and "Dev:" from the all link title %>
-              <%= admin_nav_select_link_active[:title]&.split(":").last %>
-            </div>
-          <% end %>
-        <% end %>
-        <div class="fancy-select unfancy no-restore-on-backspace">
-          <%= select(:admin, :other_navigation, options_for_select(admin_nav_select_links.map { |l| [l[:title], l[:path]] }), { prompt: admin_nav_select_prompt }, { class: "form-control" }) %>
-        </div>
+        <%= render(Admin::Combobox::Component.new(developer: current_user.developer?)) %>
       <% end %>
 
       <%= link_to "Exit Admin", root_path, class: "btn btn-outline-secondary ml-auto less-strong d-none d-lg-inline-block" %>
